fix: modal as angular component
