--- src/app/mentores/mentores/mentoria.html
@@ -1,7 +1,0 @@
-<ng-container *ngIf="show">
-  <div class="modal">
-    <ng-content></ng-content>
-  </div>
-  <div (click)="toggle()"
-       class="overlay"></div>
-</ng-container>
