--- src/app/mentores/mentores/mentoria.ts
@@ -1,18 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { ModalComponent } from './mentores.component';
-
-@NgModule({
-  declarations: [
-    ModalComponent,
-  ],
-  imports: [
-    CommonModule,
-  ],
-  exports: [
-    ModalComponent,
-  ]
-})
-export class ModalModule {
-
-}
