+++ src/app/mentores/mentores/modal/modal.component.html
@@ -1,0 +1,7 @@
+<ng-container *ngIf="show">
+    <div class="modal">
+      <ng-content></ng-content>
+    </div>
+    <div (click)="toggle()"
+         class="overlay"></div>
+  </ng-container>
